Add files via upload
OSCAR is configured in such a way that a file (image) is uploaded to a MinIO bucket, this in turn invokes two services. This is done by setting the input buckets of each service equal.
This workflow runs an example of sending the file to the input bucket and capturing the results in each output bucket of each service.